Donot add cherrypy as a default
This commit introduces changes so we don't add cherrypy support for ws4py port as we don't use it and updating ws4py to the latest version conflicts with cherrypy's version as it has not been updated in a few years. We don't use cherrypy, so we can just not build ws4py with it.