Add mac_ntpd module to build
mac_ntpd lets the NTP daemon run as a non-root user. The ntpd rc script tries
to load it, but we are missing the module so we see

kldload: can't load mac_ntpd: No such file or directory

on the console and ntpd is run as root instead of the ntpd user.

We have the MAC framework in our kernel config already, all we're missing is
the mac_ntpd module in the build profile config.

Ticket: NAS-105526